Refactoring
Taken the original 'app.py' which contained all the code, and have now split it up into multiple individual files for code readability and maintainability
